Compatibility checks: move to Utils namespace
